updated file structure of components folder
