+++ app/items/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useEffect, useState } from 'react'
-import Table from '@/app/components/Table/Table'
-import Toolbar from '@/app/components/Toolbar/Toolbar'
+import Table from '@/app/components/items-table/Table'
+import Toolbar from '@/app/components/items-table/Toolbar'
 
 export interface Item {
 	name: string
